ingestion/requirements.txt
azure-functions
azure-storage-blob>=12.19.0
requests>=2.31.0 